Add some common setup stuff
I'm sick of having to remove a couple dotfiles on most Linux installs to have my dotfiles link in properly.  There is certainly a better way of doing this, but this is the quick and dirty.
I also want a couple directories created.  Again, better ways of doing this, but this is quick and dirty.